Crate ixlist

Structs

Cursor
A cursor points to a location in a list, and you can step the cursor forward and backward.

Iter

IterMut

List
List is a doubly linked list stored in one contiguous allocation.

Node

Enums

Seek