Struct ixlist::List

pub struct List<T> {
    // some fields omitted
}

List is a doubly linked list stored in one contiguous allocation.

It is like a list implemented with pointers, except instead of pointers we use indices into a backing vector.

Features

O(1) insert and remove both at front and back. O(1) insert anywhere if you have a cursor to that position.

Can be generic over the index type (not yet implemented), so that internal prev/node links can use less space than a regular pointer (can be u16 or u32 index).

Discussion

Idea (not yet implemented): Fixate node positions at certain intervals, e.g. every 32nd node is always in its correct index in the backing vector??

With some cleanup we can use unchecked indexing for impl.

Methods

impl<T> List<T>

fn new() -> List<T>

Create a new List.

fn with_capacity(cap: usize) -> List<T>

Create a new List with specified capacity.

fn len(&self) -> usize

fn iter(&self) -> Iter<T>

fn iter_mut(&mut self) -> IterMut<T>

fn cursor(&mut self) -> Cursor<T>

fn push_front(&mut self, value: T)

fn push_back(&mut self, value: T)

fn pop_front(&mut self) -> Option<T>

fn pop_back(&mut self) -> Option<T>

fn linearize(&mut self)

Reorder internal datastructure into traversal order